+++ PopcornPals-Server/src/main/java/bg/fmi/popcornpals/model/Studio.java
@@ -27,7 +27,7 @@ public class Studio {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer ID;
+    private Long ID;
 
     @NotNull(message = "Studio: Name cannot be null")
     @NotBlank(message = "Studio: Name cannot be blank")
